fix: lint backend files, .js, .ts, and .tsx
It might seem weird to change "tsconfig.json" to get eslint
to work, but we have:

"parser": "@typescript-eslint/parser" in ".eslintrc", which means
that TypeScript is responsible for doing all the parsing for eslint.

And since we also specified:

"project": ["./tsconfig.json"] in ".eslintrc", which is a good
and standard thing to do to have eslint's rules fully exploit
the power of TypeScript, we have, I think, no choice other than
include all the files we want to lint into the tsconfig file,
and it actually sort of makes sense.